Add weird parsing thing to parse into specific custom attributes
Will need to be cleaned up but wanted to commit it so I didn't lose it
