+++ emulation_system/emulation_system/compose_file_creator/input/models/container_types/base_type.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class HardwareSpecificAttributes(BaseModel):
+
+    class Config:
+        extra = "forbid"
+
