+++ emulation_system/emulation_system/compose_file_creator/input/models/container_types/heater_shaker_module.py
@@ -1,14 +1,8 @@
-from enum import Enum
-from typing_extensions import Literal
-from compose_file_creator.input.models.container_types.base_attributes import (
-    BaseAttributes
-)
-from compose_file_creator.input.settings import (
-    Hardware,
+from emulation_system.compose_file_creator.input.models.container_types.base_type import HardwareSpecificAttributes
+from emulation_system.compose_file_creator.input.settings import (
     HeaterShakerModes
 )
 
 
-class HeaterShakerModuleAttributes(BaseAttributes):
-    hardware: Literal[Hardware.HEATER_SHAKER_MODULE.value]
+class HeaterShakerModuleAttributes(HardwareSpecificAttributes):
     mode: HeaterShakerModes = HeaterShakerModes.SOCKET
