+++ emulation_system/emulation_system/compose_file_creator/input/models/container_types/thermocycler_module.py
@@ -1,14 +1,10 @@
 from pydantic import Field
-
-from typing_extensions import Literal
-from compose_file_creator.input.models.container_types.base_attributes import (
-    BaseAttributes
+from emulation_system.compose_file_creator.input.models.container_types.base_type import HardwareSpecificAttributes
+from emulation_system.compose_file_creator.input.settings import (
+    TemperatureModelSettings
 )
-from compose_file_creator.input.settings import Hardware, TemperatureModelSettings
 
-
-class ThermocyclerModuleAttributes(BaseAttributes):
-    hardware: Literal[Hardware.THERMOCYCLER_MODULE.value]
+class ThermocyclerModuleAttributes(HardwareSpecificAttributes):
     lid_temperature: TemperatureModelSettings = Field(
         alias="lid-temperature", default=TemperatureModelSettings()
     )
